UI修正

+++ Game/Game/source/Boss.cpp
@@ -588,8 +588,8 @@ void Boss::Render()
 {
 	DrawGraph(300, 900, iconhandle, true);
 	DrawGraph(300, 950, flamehandle, true);
-	int HPgauge = (1185 / BossMaxHP) * BossHP;
-	DrawRectGraph(310, 960, 0, 0,HPgauge, 30, HPhandle, true, false);
+	float HPgauge =  (1185.0f / BossMaxHP) * BossHP;
+	DrawRectGraph(310, 960, 0, 0, static_cast<int>(HPgauge), 30, HPhandle, true, false);
 
 	{
 		if (!downFlag) {
